The ClientInterface have moved into the interface directory.
All includes are fixed to new namespace.
