+++ Source/Client/Response.php
@@ -8,7 +8,7 @@ use InvalidArgumentException;
 use Magento\Framework\App\ObjectManager;
 use PagoFacil\Payment\Exceptions\HttpException;
 use PagoFacil\Payment\Exceptions\PaymentException;
-use PagoFacil\Payment\Source\Client\ClientInterface as HTTPInterface;
+use PagoFacil\Payment\Source\Client\Interfaces\ClientInterface as HTTPInterface;
 use PagoFacil\Payment\Source\Client\Interfaces\PagoFacilResponseInterface;
 use PagoFacil\Payment\Source\Interfaces\Dto;
 use PagoFacil\Payment\Source\Transaction\Charge;
